Replace team types
