+++ lib/team-client/index.js
@@ -96,7 +96,7 @@ var Team = /** @class */ (function (_super) {
      */
     Team.prototype.removeMember = function (removal) {
         return __awaiter(this, void 0, void 0, function () {
-            var options, res;
+            var options;
             return __generator(this, function (_a) {
                 switch (_a.label) {
                     case 0: return [4 /*yield*/, this._guardInitialized()];
@@ -105,8 +105,8 @@ var Team = /** @class */ (function (_super) {
                         options = removal;
                         return [4 /*yield*/, this._runApiCommand({ apiName: 'team', method: 'remove-member', options: options })];
                     case 2:
-                        res = _a.sent();
-                        return [2 /*return*/, res];
+                        _a.sent();
+                        return [2 /*return*/];
                 }
             });
         });
